image_organise/requirements.txt
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.1.0
Werkzeug==0.14.1
click==6.7
itsdangerous==0.24
